Turn and Play class changed
In Play class: added the Move() method. When the Move() method is
called, the current player switches in the Turn class.

In Turn class: Determines based on calls from Move() method whose turn
it is, also determines the total number of turns taken. Use
getCurrPlayer() to figure out whose turn it is, use getCurrTurn() to
determine the total number of turns.