Significant Integration Changes
I have significantly changed the functionality of Game, Player, Seed,
and Row. I have implemented a singleton functionality for a lot of
classes that need to have only one single instantiation. Please use the
.getInstance() method to instantiate these classes. This includes Turn,
Rules, and GameBoard, though we will probably make singletons out of
HelpDoc, BoardTypes, DrawingScreen, and MainWindow. The cool part: it is
now possible to make basic moves on a text based version of Wari. Still
waiting on GUI integration and scoring functionality.