Started work on incorporating GUI
Added rules text on the side and a 2x6 GridLayout of buttons to
represent each hole on the Wari board. The buttons have text titles to
represent the amount of seeds in them. Next objective: add functionality
to the buttons. When a button is clicked it should move the seeds from
that hole and switch turns. After that, we can start working on bringing
in the seed / hand images we were provided.