Starting Work on Computer / Human Player Choice
You can decide to play against either a human or computer opponent at
the beginning of the GUI. No computer player functionality yet. Thinking
about making a really simple computer opponent that just picks the hole
with the most seeds in it every time. Wouldn't be a good opponent, but
it would be an easy solution we could throw together and iron out later
if we need a quick and dirty implementation.